Improve naming consistency
