+++ Demo/Sources/Views/Player/ScrubbingSpeed.swift
@@ -6,52 +6,30 @@
 
 import Foundation
 
-enum ScrubbingSpeed: HSliderScrubbingSpeed, Equatable {
-    case hiSpeed
-    case halfSpeed
-    case quarterSpeed
-    case fine
-
+struct StandardScrubbingSpeed: ScrubbingSpeed {
     static var `default`: Self {
-        .hiSpeed
+        .init()
     }
 
     var value: Double {
-        switch self {
-        case .hiSpeed:
-            return 1
-        case .halfSpeed:
-            return 0.5
-        case .quarterSpeed:
-            return 0.25
-        case .fine:
-            return 0.1
-        }
+        1
     }
 
-    var name: LocalizedStringResource {
-        switch self {
-        case .hiSpeed:
-            return "High-Speed Scrubbing"
-        case .halfSpeed:
-            return "Half-Speed Scrubbing"
-        case .quarterSpeed:
-            return "Quarter-Speed Scrubbing"
-        case .fine:
-            return "Fine Scrubbing"
-        }
-    }
+    private init() {}
 
     static func speed(forDistance distance: CGFloat) -> Self {
-        switch distance {
-        case 0..<50:
-            return .hiSpeed
-        case 50..<100:
-            return .halfSpeed
-        case 100..<150:
-            return .quarterSpeed
-        default:
-            return .fine
-        }
+        .default
     }
+}
+
+/// Defines a scrubbing speed.
+protocol ScrubbingSpeed {
+    /// The default scrubbing speed.
+    static var `default`: Self { get }
+
+    /// The (positive) associated numeric value.
+    var value: Double { get }
+
+    /// The speed to apply at a given distance from the slider.
+    static func speed(forDistance distance: CGFloat) -> Self
 }
