+++ Demo/Sources/Views/Player/TimeBar~ios.swift
@@ -12,7 +12,7 @@ struct TimeBar: View {
     @ObservedObject var player: Player
     @ObservedObject var visibilityTracker: VisibilityTracker
     @Binding var isInteracting: Bool
-    @Binding var scrubbingSpeed: ScrubbingSpeed
+    @Binding var scrubbing: Scrubbing
 
     @StateObject private var progressTracker = ProgressTracker(
         interval: CMTime(value: 1, timescale: 10),
@@ -20,7 +20,7 @@ struct TimeBar: View {
     )
 
     var body: some View {
-        TimeSlider(player: player, progressTracker: progressTracker, visibilityTracker: visibilityTracker, scrubbingSpeed: $scrubbingSpeed)
+        TimeSlider(player: player, progressTracker: progressTracker, visibilityTracker: visibilityTracker, scrubbing: $scrubbing)
             .onChange(of: progressTracker.isInteracting) { isInteracting = $0 }
             .bind(progressTracker, to: player)
     }
